refactor explorer to renderer file controls to codebook using a callback
