+++ src/explorer/layout.js
@@ -3,7 +3,5 @@
 \------------------------------------------------------------------------------------------------*/
 
 export function layout() {
-  this.controls.wrap = this.wrap.append('div').attr('class', 'controls');
-
   this.codebookWrap = this.wrap.append('div').attr('class', 'codebookWrap');
 }
